refactor: split alt-text service tests
