+++ apps/location-manager/packages/python-alt-text/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the Vertex content service."""
